Conflicts:
	Core/Tools/inc/MagneticCoefficientsMap.h

+++ Core/Tools/inc/OutputData.h
@@ -370,7 +370,7 @@ typename OutputData<T>::const_iterator OutputData<T>::begin() const
 template <class T>
 typename OutputData<T>::iterator OutputData<T>::begin(ThreadInfo thread_info)
 {
-    size_t total_size = getAllocatedSize();
+    int total_size = (int)getAllocatedSize();
     int start_index = thread_info.getBeginIndex(total_size);
 
     typename OutputData<T>::iterator result(this, start_index);
@@ -383,7 +383,7 @@ template <class T>
 typename OutputData<T>::const_iterator
     OutputData<T>::begin(ThreadInfo thread_info) const
 {
-    size_t total_size = getAllocatedSize();
+    int total_size = (int)getAllocatedSize();
     int start_index = thread_info.getBeginIndex(total_size);
 
     typename OutputData<T>::const_iterator result(this, start_index);
@@ -395,7 +395,7 @@ typename OutputData<T>::const_iterator
 template <class T>
 typename OutputData<T>::iterator OutputData<T>::end(ThreadInfo thread_info)
 {
-    size_t total_size = getAllocatedSize();
+    int total_size = (int)getAllocatedSize();
     int end_index = thread_info.getEndIndex(total_size);
 
     typename OutputData<T>::iterator result(this, end_index);
@@ -408,7 +408,7 @@ template <class T>
 typename OutputData<T>::const_iterator
     OutputData<T>::end(ThreadInfo thread_info) const
 {
-    size_t total_size = getAllocatedSize();
+    int total_size = (int)getAllocatedSize();
     int end_index = thread_info.getEndIndex(total_size);
 
     typename OutputData<T>::const_iterator result(this, end_index);
